+++ packages/jest-util/src/deep_cyclic_copy.js
@@ -14,7 +14,7 @@ export type DeepCyclicCopyOptions = {|
   keepPrototype: boolean,
 |};
 
-// Node 6 does not have gOPDs, so we define a simple polyfill for it.
+// $FlowFixMe: Node 6 does not have gOPDs, so we define a simple polyfill for it.
 if (!Object.getOwnPropertyDescriptors) {
   // $FlowFixMe: polyfill
   Object.getOwnPropertyDescriptors = obj => {