+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 * `[jest-runtime]` Allow for transform plugins to skip the definition process
   method if createTransformer method was defined.
   ([#5999](https://github.com/facebook/jest/pull/5999))
+* `[expect]` Add stack trace for async errors
+  ([#6008](https://github.com/facebook/jest/pull/6008))
+* `[jest-jasmine2]` Add stack trace for timeouts
+  ([#6008](https://github.com/facebook/jest/pull/6008))
+* `[jest-jasmine2]` Add stack trace for thrown non-`Error`s
+  ([#6008](https://github.com/facebook/jest/pull/6008))
 * `[jest-runtime]` Prevent modules from marking themselves as their own parent
   ([#5235](https://github.com/facebook/jest/issues/5235))
 * `[jest-mock]` Add support for auto-mocking generator functions